doctorAPI/requirements.txt
Django==4.0.5
django-filter==22.1
djangorestframework==3.13.1
drf-yasg==1.20.0
mysqlclient==2.1.1
